requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
django-ninja==1.3.0
dnspython==2.7.0
email_validator==2.2.0
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
ruamel.base==1.0.0
sets==0.3.2
uvicorn==0.34.0
whitenoise==6.9.0